document zero-indexed lists, catch invalid cycles on construction
The function `permutation(list)` uses zero-indexed lists, because that's
how lists work in JME. The documentation now mentions this.

Previously, permutations weren't checked for validity on construction,
so errors would only be found once you do something which requires
computing cycles.
This changes the constructor so it computes the cycles immediately. It
also catches invalid cycles which go outside the range of the given map,
or include duplicate elements.

see numbas/Numbas#793