throw an error if trying to make an element bigger than 10^6.
fixes #1

I considered changing the internal representation to something like a
dictionary, to allow working with a small subgroup of a much bigger
permutation group (i.e. the cycle representation is short, but the
individual elements have very big indices), but that'd be a lot of
effort for something that nobody is practically ever going to want to
do.

So this change makes it throw an error if you try to use an element
bigger than one million when constructing a permutation from cycle
notation, and suggests that you might have forgotten to put commas
between elements.